String and Literal Parsing Complete

+++ src/token_handler/error_handler.rs
@@ -1,13 +1,13 @@
 
 static mut HAS_ERROR:bool = false;
-pub fn error(line:i32, message:String )
+pub fn error(line:usize, message:String )
 {
     report(line, "".to_string(), message);
 }
 
 
 
-fn report(line:i32, location:String, message:String)
+fn report(line:usize, location:String, message:String)
 {
     eprintln!("Line {line_no} Error {whe} : {mes}", line_no=line, whe=location, mes=message);
